fix(060): wire public bundle to new routes + hash mode for dev
Root cause: web/src/routes/status-routes.ts was created but never imported.
The actual entry point (status-main.ts → router/status-router.ts) still
held the old /status, /status/:id routes pointing at deleted views, so
the public bundle was unreachable. The admin bundle's leftover /status
route also broke after the API shape change.

- router/status-router.ts: 4 routes (/, /history, /uptime, /resource/:id)
  with hash history → /status.html#/uptime works in dev without server
  SPA fallback, and any custom domain works in prod.
- router/index.ts (admin): drop /status + /status/:id legacy routes; the
  public bundle is the only owner now.
- routes/status-routes.ts: deleted (duplicate).
- status-router.spec.ts: rewritten for the 4 spec 060 routes.

Dev URLs:
  http://localhost:5173/status.html        → current snapshot
  http://localhost:5173/status.html#/history
  http://localhost:5173/status.html#/uptime
